pyproject.toml
[tool.poetry]
name = "web-rag-scraper"
version = "0.1.0"
description = "Production grade Web Scraper + Knowledge Graph RAG System"
authors = ["Akhilesh-Mohanasundaram Kaniska"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"

# --- API & Core ---
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"

# --- Async & Tasks ---
celery = {extras = ["redis"], version = "^5.3.6"}
redis = "^5.0.3"
httpx = "^0.28.1"

# --- Database ---
neo4j = "^5.18.0"
sqlalchemy = "^2.0.28"
asyncpg = "^0.29.0"

# --- AI & RAG (LlamaIndex Stack) ---
llama-index = "^0.14.10"
llama-index-graph-stores-neo4j = "^0.5.1"
llama-index-llms-openai = "^0.6.12"
llama-index-embeddings-openai = "^0.5.1"

# --- Scraping ---
playwright = "^1.42.0"
beautifulsoup4 = "^4.12.3"
trafilatura = "^1.8.0" # Excellent for extracting main text from messy HTML
asgiref = "^3.11.0"
llama-index-core = "^0.14.10"
llama-index-llms-gemini = "^0.6.1"
llama-index-embeddings-gemini = "^0.4.1"
tenacity = "^9.1.2"
streamlit = "^1.52.2"
requests = "^2.32.5"
google-genai = "^1.56.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"